normalize image name
