+++ package.json
@@ -37,7 +37,7 @@
         "onCommand:vscode-docker.compose.restart",
         "onCommand:vscode-docker.compose.up",
         "onCommand:vscode-docker.configure",
-        "onCommand:vscode-docker.configureDocker",
+        "onCommand:vscode-docker.configureCompose",
         "onCommand:vscode-docker.containers.attachShell",
         "onCommand:vscode-docker.containers.browse",
         "onCommand:vscode-docker.containers.configureExplorer",
@@ -1991,8 +1991,8 @@
                 "category": "%vscode-docker.commands.category.docker%"
             },
             {
-                "command": "vscode-docker.configureDocker",
-                "title": "%vscode-docker.commands.configureDocker%",
+                "command": "vscode-docker.configureCompose",
+                "title": "%vscode-docker.commands.configureCompose%",
                 "category": "%vscode-docker.commands.category.docker%"
             },
             {
@@ -2444,9 +2444,9 @@
         "@types/string-replace-webpack-plugin": "^0.1.0",
         "@types/tar": "^4.0.3",
         "@types/xml2js": "^0.4.5",
-        "@typescript-eslint/eslint-plugin": "^2.22.0",
-        "@typescript-eslint/eslint-plugin-tslint": "^2.22.0",
-        "@typescript-eslint/parser": "^2.22.0",
+        "@typescript-eslint/eslint-plugin": "^2.23.0",
+        "@typescript-eslint/eslint-plugin-tslint": "^2.23.0",
+        "@typescript-eslint/parser": "^2.23.0",
         "adm-zip": "^0.4.14",
         "copy-webpack-plugin": "^5.1.1",
         "eslint": "^6.8.0",
